fix: move error and its type to exceptions directory
