+++ src/exceptions/error-type.ts
